requirements/base.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
certifi==2022.9.14
chardet==5.0.0
coreapi==2.3.3
coreschema==0.0.4
cycler==0.10.0
Django==4.1
django-appconf==1.0.5
django-compressor==4.1
django-easy-audit==1.3.3
django-extensions==3.2.1
django-filter==22.1
django-fixture-magic==0.1.5
django-guardian==2.4.0
django-rest-swagger==2.2.0
djangorestframework==3.13.1
djangorestframework-bulk==0.2.1
djangorestframework-csv==2.1.1
djangorestframework-gis==1.0
drf-compound-fields==2.0.0
drf-extensions==0.7.1
fuzzywuzzy==0.18.0
geojson==2.5.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.6.0
mpld3==0.5.8
numpy==1.23.3
olefile==0.46
openapi-codec==1.3.2
pandas==1.4.4
pdfkit==1.0.0
Pillow==9.2.0
pipupgrade==1.12.0
psycopg2==2.9.4
pyparsing==3.0.9
PyPDF2==2.10.8
python-dateutil==2.8.2
python-Levenshtein==0.12.2
pytz==2022.2.1
rcssmin==1.1.0
requests==2.28.1
rest-condition==1.0.3
rjsmin==1.2.0
scipy==1.9.1
seaborn==0.12.0
simplejson==3.17.6
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
unicodecsv==0.14.1
uritemplate==4.1.1
urllib3==1.26.12
xlrd==2.0.1
XlsxWriter==3.0.3
xlwt==1.3.0
gunicorn==20.1.0

